+++ docs/CHANGELOG.md
@@ -3,7 +3,11 @@
 
  - new 'toggle annotation sidebar' for the viewer mode.
  
- -    
+ - manual updates might be set
+ 
+ - more GA changes so I can track and improve polar usage
+ 
+ -     
 
 # 1.0.0-beta200